Lift the repair worker endpoints out of web_server.py
26 routes to api/repair.py (-460 lines): status, job control, findings, history.
The worker singleton stays boot-owned by web_server (progress callbacks + the
automation event bridge); the module reads it through a getter since it is
legitimately None when its init failed. Verified by url-map diff (1,152 routes,
method+path identical) and an in-app smoke of status + findings.
